requirements.txt
backports.ssl-match-hostname==3.5.0.1
apache-libcloud==1.5.0
python-cinderclient==1.11.0
keystoneauth1==2.18.0
python-novaclient==8.0.0
aliyun-python-sdk-core==2.3.1
